# MAINT: Speed up np.maximum/np.minimum by up to 15x when operating on two inputs