@eric-wieser 2019-03-29
I adjusted the title - this specifically targets `np.maximum` not `np.max == np.maximum.reduce`