@mattip 2019-04-18
Using NPY_GCC_OPT_3 seems like a win. There is a [remark](https://github.com/numpy/numpy/blob/master/numpy/core/src/umath/fast_loop_macros.h#L92) about not using it too much since it can lead to code bloat. Any opinions on the cost of it here?