@qwhelan 2019-04-18
I can do an actual comparison sometime tomorrow, but a similar change to `np.abs()` increased `loops.o` by `0.16%` (https://github.com/numpy/numpy/pull/13271). This will likely be a bit more due to more types and logic.