@qwhelan 2019-04-22
Unfortunately, `BINARY_LOOP_FAST` causes quite a bit more bloat (`~7%`) than the example I gave. Most of the macro is dealing with potential aliasing and scalar comparison cases; the bloat could be cut in half or more if those cases are dropped for these functions. That being said, this increase seems a bit much for `np.maximum` given that the reduction version is far more common.