@mattip 2019-04-22
On my machine (Ubuntu 18.04 x64) the so size grows by 2.4% to 16MB.