@juliantaylor 2019-04-24
the fast loop was not applied to many of these because it seemed like unnecessary bloat for little used functions or types

Though I have no issue with bloating a bit for faster integer and float operations, in the end it is mostly just size on disk.
datetime and half floats imo do not need the fast loop but then again those are only three types compared to the many different integers so it probably makes no difference in bloat size and avoids having to look at them again in similar changes in the future.