@mattip 2019-04-25
So sounds like this PR is almost ready to go. @qwhelan could you address the indent and exclude datetime/float16?