@mattip 2019-04-29
LGTM. I guess this should get a mention in the Improvement section of the release notes, we have already noted `numpy.exp` and `numpy.log` so putting it next to that seems appropriate.