@rgommers 2019-04-29
Can we conclude on the bloat discussion first? Can one of you summarize the current state?