@mattip 2019-04-29
On my machine using gcc-7 the change from the commit just before this one to  8b5196d is 
17192648 bytes -> 17611176 bytes, a change of 408kB out of 16789kB for around 2.4%. The mailing list discussion did not provide a hard criteria for a decision, although the general direction was that disk size does not matter as much as download size.