@rgommers 2019-04-30
Thanks @mattip. My take on this is:

- This case is borderline. If we would write down a hard criterion, this likely would not meet it.
- Rationale: if we get 100 PRs like this, the average performance of numpy for a user would not change much, however we have by then blown up the size NumPy takes up (disk/RAM/download/etc) by a factor 2.4
- *However*, we won't get 100 PRs like this. So judging this based on such a criterion isn't quite right.
- We have this PR now, and it's good to go. Presumably it helps @qwhelan significantly. So I'm +0.5 for merging it.
- Also note that Cython has the same problem: taking one function and putting it in a .pyx file gives a huge amount of bloat (example: `scipy.ndimage.label`). We had the same discussion there, but it never became a practical issue because there were not many other PRs like that.

tl;dr let's merge this, and let's try not to make these kinds of changes a habit