@charris 2022-06-10
@qwhelan Needs a rebase. I'm wondering if these functions have been vectorized (SIMD). @seberg Do you recall?