@qwhelan 2022-06-10
I won't have bandwidth to do that, so anyone interested should feel free to
do so.

On Thu, Jun 9, 2022, 5:26 PM Charles Harris ***@***.***>
wrote:

> @qwhelan <https://github.com/qwhelan> Needs a rebase. I'm wondering if
> these functions have been vectorized (SIMD). @seberg
> <https://github.com/seberg> Do you recall?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/numpy/numpy/pull/13207#issuecomment-1151744504>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AADLOH77VZ52YNATMJFG76DVOKDTJANCNFSM4HCGFX5A>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>