@seberg 2022-06-10
Yeah, we have a whole file for it here: https://github.com/numpy/numpy/blob/49c560c22f137907ea6a240591e49b004f28444b/numpy/core/src/umath/loops_minmax.dispatch.c.src

That said, I guess there are a couple of places here that could still use the attribute, for example the datetime (and maybe half) loops.

There may be a general question whether we should be adding such optimization attributes basically by default to most things in `loops.c.src`?